Fix migration test for Python 3.11
- Under python 3.11, pkgutil.resolve_name(), called by
unittest.mock.patch, will raise an ValueError if the name to be resolved
(the `target` argument to patch()) starts with a number. So we need to 1)
refer to the function to be patched indirectly via the current module and 2) rename the test
file itself so it no longer contains a number in its name.
- More info on https://github.com/python/cpython/issues/100950
- Check migrations for linting and formatting